refactor(shortcuts): inline recording, pencil icon, fix Mac Ctrl→mod bug
- Replace fullscreen ShortcutRecordingModal with inline InlineRecorder that
  captures keys directly in the chip row — no backdrop, no modal
- Live partial preview: modifier-only presses show a pulsing partial combo
  (e.g. Ctrl… then Ctrl+Shift+K) before the final key is pressed
- Pencil icon appears on chip hover; click chip or pencil to edit inline;
  click outside or Escape to cancel
- Remove Keycap press animation — chips are purely visual, no confusing
  press effect on non-editable shortcuts
- Fix eventToCombo Mac bug: Ctrl key on Mac was mapped to mod (⌘) instead
  of ctrl (⌃); same fix in tiptapEventToCombo and eventToSearchCombo
- Allow shift+non-printable combos (shift+up/down) in the recorder so
  prompt-history defaults can be re-recorded; block shift+letter to avoid
  conflicts with normal typing
- Move ShortcutRecorder from primitives/ into features/command/
- Delete dead primitives/KeyboardShortcutsSheet.tsx
- Fix Windows encoding corruption in useTiptapEditor.ts comment
